+++ src/classify/classify_policy.rs
@@ -1,5 +1,5 @@
 use crate::classify::{classification_code::*, query_ptir::QueryPTIR, ref_ptir::RefPTIRManager};
 
-pub fn run_classify(query_ptir: &QueryPTIR, ref_ptir: &RefPTIRManager) -> ClassCode {
+pub fn run_classify(_query_ptir: &QueryPTIR, _ref_ptir: &RefPTIRManager) -> ClassCode {
     todo!()
 }