+++ src/constants.rs
@@ -1,1 +1,3 @@
 pub const ISOM_GTF_SCHEMA: &str = "1";
+pub const ISOMX_VERSION: u8 = 1;
+pub const ISOMS_VERSION: u8 = 1;