+++ src/core/tx_type.rs
@@ -1,6 +1,7 @@
 use core::fmt;
 
 #[derive(Copy, Clone, Debug)]
+#[repr(u8)]
 pub enum TxType {
     MONO,
     ALLC,
@@ -15,6 +16,20 @@ impl TxType {
             "ALL_CA" => Some(TxType::ALLC),
             "PRT_CA" => Some(TxType::PRTC),
             "NOT_CA" => Some(TxType::NOTC),
+            _ => None,
+        }
+    }
+
+    pub fn to_u8(self) -> u8 {
+        self as u8
+    }
+
+    pub fn from_u8(v: u8) -> Option<TxType> {
+        match v {
+            0 => Some(TxType::MONO),
+            1 => Some(TxType::ALLC),
+            2 => Some(TxType::PRTC),
+            3 => Some(TxType::NOTC),
             _ => None,
         }
     }